requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
slack_sdk==3.33.1
sqlparse==0.5.1
